+++ packages/collapse/tests/collapse.test.tsx
@@ -1,9 +1,10 @@
 import { Collapse } from "../src"
 import { render, screen } from "@testing-library/react"
 import "@testing-library/jest-dom"
-import React from "react"
 import { CollapseProps } from "../src"
 import { globalColor, illaPrefix } from "@illa-design/theme"
+import { ShareIcon } from "@illa-design/icon"
+import userEvent from "@testing-library/user-event"
 
 function DemoTest(props: CollapseProps) {
   const CollapseItem = Collapse.Item
@@ -16,9 +17,10 @@ function DemoTest(props: CollapseProps) {
       <CollapseItem
         header="Life-was-like-a-box-of-chocolates-you-never-know-what-your-gonna-get"
         name="1"
+        extra={<ShareIcon />}
         data-testid={"collapse-item-1"}
       >
-        Life-was-like-a-box-of-chocolates-you-never-know-what-your-gonna-get
+        First
       </CollapseItem>
 
       <CollapseItem
@@ -27,7 +29,7 @@ function DemoTest(props: CollapseProps) {
         disabled
         data-testid={"collapse-item-2"}
       >
-        Life-was-like-a-box-of-chocolates-you-never-know-what-your-gonna-get
+        Second
       </CollapseItem>
 
       <CollapseItem
@@ -35,7 +37,7 @@ function DemoTest(props: CollapseProps) {
         name="3"
         data-testid={"collapse-item-3"}
       >
-        Life-was-like-a-box-of-chocolates-you-never-know-what-your-gonna-get
+        Third
       </CollapseItem>
     </Collapse>
   )
@@ -46,4 +48,12 @@ test("Collapse renders with border", () => {
   expect(screen.getByTestId("collapse-wrapper")).toHaveStyle({
     border: `border: solid 1px ${globalColor(`--${illaPrefix}-gray-08`)}`,
   })
+})
+
+test("Collapse renders with extra icon", () => {
+  render(<DemoTest defaultActiveKey={["1", "2"]} />)
+  const icon = screen.getByTitle("ShareIcon")
+  expect(icon).toBeInTheDocument()
+  userEvent.click(icon)
+  expect(screen.getByTestId("collapse-item-1").lastChild).toBeVisible()
 })