+++ packages/modal/stories/modal.stories.tsx
@@ -6,7 +6,6 @@ import { Button } from "@illa-design/button"
 import { Input } from "@illa-design/input"
 import { Message } from "@illa-design/message"
 import { InfoCircleIcon, SuccessIcon } from "@illa-design/icon"
-
 import { ConfigProvider, zhCN, enUS } from "@illa-design/config-provider"
 
 export default {
@@ -53,8 +52,6 @@ const Template: Story<ModalProps> = (args) => {
           confirmLoading={confirmLoading}
           onCancel={() => setVisible(false)}
           onOk={onOk}
-          afterClose={() => console.log("close")}
-          afterOpen={() => console.log("open")}
         >
           I don't know if we each have a destiny, or if we're all just floating
           around accidentally―like on a breeze.