+++ packages/tree-select/tests/tree-select.test.tsx
@@ -63,6 +63,25 @@ const data = [
 ]
 
 test("TreeSelect renders correctly ", () => {
-  render(<TreeSelect data-testid={"TreeSelect"} treeData={data} />)
+  render(<TreeSelect data-testid="TreeSelect" treeData={data} />)
   expect(screen.getByTestId("TreeSelect")).toBeInTheDocument()
+})
+
+test("TreeSelect renders with defaultValue ", () => {
+  render(
+    <TreeSelect data-testid="TreeSelect" defaultValue="0-0" treeData={data} />,
+  )
+  expect(screen.getByText("0-0-head")).toBeInTheDocument()
+})
+
+test("TreeSelect renders with multiple defaultValue ", () => {
+  render(
+    <TreeSelect
+      data-testid="TreeSelect"
+      defaultValue={["0-0"]}
+      multiple
+      treeData={data}
+    />,
+  )
+  expect(screen.getByText("0-0-head")).toBeInTheDocument()
 })