+++ packages/comment/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @illa-design/comment
 
+## 0.0.2
+
+### Patch Changes
+
+- update all components version
+
+- Updated dependencies []:
+  - @illa-design/system@0.0.2
+  - @illa-design/theme@0.0.2
+
 ## 0.0.1
 
 ### Patch Changes