+++ packages/pagination/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @illa-design/pagination
 
+## 0.0.2
+
+### Patch Changes
+
+- update all components version
+
+- Updated dependencies []:
+  - @illa-design/icon@0.0.2
+  - @illa-design/input@0.0.2
+  - @illa-design/select@0.0.2
+  - @illa-design/system@0.0.2
+  - @illa-design/theme@0.0.2
+
 ## 0.0.1
 
 ### Patch Changes