+++ packages/select/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @illa-design/select
 
+## 0.0.2
+
+### Patch Changes
+
+- update all components version
+
+- Updated dependencies []:
+  - @illa-design/checkbox@0.0.2
+  - @illa-design/empty@0.0.2
+  - @illa-design/icon@0.0.2
+  - @illa-design/input@0.0.2
+  - @illa-design/input-tag@0.0.2
+  - @illa-design/list@0.0.2
+  - @illa-design/system@0.0.2
+  - @illa-design/theme@0.0.2
+  - @illa-design/trigger@0.0.2
+
 ## 0.0.1
 
 ### Patch Changes