+++ packages/system/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @illa-design/system
 
+## 0.0.2
+
+### Patch Changes
+
+- update all components version
+
 ## 0.0.1
 
 ### Patch Changes